refactor(button): consolidate Button component implementations
Merge Button v2 into main Button component and remove duplicate code.
This change simplifies the component API by removing complex features
like loading states, prefix/suffix elements while maintaining core
functionality with variant and size options.

Changes:
- Remove components/shared/Button/v2/ directory
- Update Button implementation to use BoxFancy-based design
- Simplify component API with focused variant and size props
- Update all component imports across the codebase
- Remove loading, prefix/suffix functionality for cleaner interface
- Update tests and Storybook stories to match new implementation

BREAKING CHANGE: Button component no longer supports loading,
prefix, suffix, or component polymorphism props.
