+++ components/shared/Button/index.tsx
@@ -1,5 +1,3 @@
-import { Button } from "./Button";
-
 export * from "./Button";
 
-export default Button;
+export { default } from "./Button";
