--- components/shared/Button/v2/index.tsx
@@ -1,5 +1,0 @@
-import { Button as ButtonV2 } from "./Button";
-
-export * from "./Button";
-
-export default ButtonV2;
